tts/requirements.txt
fastapi
uvicorn
pydantic
transformers
torchaudio
datasets
soundfile
sentencepiece